+++ frontend/src/App.jsx
@@ -17,22 +17,24 @@ function App() {
       <Routes>
         <Route path="/" element={<Layout />}>
           <Route index element={<Home />} />
-          <Route
+          {/* <Route
             path="/recipe"
             element={
               <PrivateRoute>
                 <Recipe />
               </PrivateRoute>
             }
-          />
-          <Route
+          /> */}
+          {/* <Route
             path="/recipe/add"
             element={
               <PrivateRoute>
                 <AddRecipe />
               </PrivateRoute>
             }
-          />
+          /> */}
+          <Route path="/recipe" element={<Recipe />} />
+          <Route path="/recipe/add" element={<AddRecipe />} />
           <Route path="/user/" element={<Profile />} />
           {/* <Route
             path="/user/:id"
