+++ frontend/src/data/testimonials.js
@@ -1,0 +1,16 @@
+const testimonials = [
+  {
+    text: 'Thank you for the wonderful recipe for feta pasta with tomatoes and basil. It turned out to be not only tasty, but also incredibly colorful. This has become a favorite family meal!',
+    author: 'LARRY PAGEIM',
+  },
+  {
+    text: 'Amazing experience! The flavors were perfect and it was so easy to make.',
+    author: 'JANE DOE',
+  },
+  {
+    text: 'My kids loved it! And it’s now our go-to recipe for family dinners.',
+    author: 'JOHN SMITH',
+  },
+];
+
+export default testimonials;
