+++ frontend/src/pages/AddRecipe/AddRecipePage.jsx
@@ -1,11 +1,10 @@
 // src/pages/AddRecipePage/AddRecipePage.jsx
 
-import React from 'react';
 import { useNavigate } from 'react-router-dom';
 
 import PathInfo from '../../components/ui/PathInfo/PathInfo';
 import MainTitle from '../../components/ui/MainTitle/MainTitle';
-import Subtitle from '../../components/ui/Subtitle/Subtitle';
+import SubTitle from '../../components/ui/SubTitle/SubTitle';
 import AddRecipeForm from '../../components/AddRecipeForm/AddRecipeForm';
 
 const AddRecipePage = () => {
@@ -19,7 +18,7 @@ const AddRecipePage = () => {
     <div className="container">
       <PathInfo current="ADD RECIPE" onHomeClick={handleHome} />
       <MainTitle text="Add recipe" />
-      <Subtitle text="Reveal your culinary art, share your favorite recipe and create gastronomic masterpieces with us." />
+      <SubTitle text="Reveal your culinary art, share your favorite recipe and create gastronomic masterpieces with us." />
       <AddRecipeForm />
     </div>
   );
@@ -42,7 +41,7 @@ export default AddRecipePage;
 //   };
 
 //   const [text, setText] = React.useState('');
-  
+
 //   return (
 //     <div className={`container`}>
 //       {/* Хедер з заголовком */}
