+++ frontend/src/pages/Profile/Profile.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useModal } from '../../hooks/useModal';
 import PathInfo from '../../components/ui/PathInfo/PathInfo';
 import MainTitle from '../../components/ui/MainTitle/MainTitle';
-import Subtitle from '../../components/ui/Subtitle/Subtitle';
+import SubTitle from '../../components/ui/SubTitle/SubTitle';
 import Button from '../../components/Button/Button';
 import RecipePreview from '../../components/RecipePreview/RecipePreview';
 import UserInfo from '../../components/UserInfo/UserInfo';
@@ -14,9 +14,9 @@ const Profile = () => {
     <div class="container">
       <PathInfo current="Profile" />
       <MainTitle text="Profile" />
-      <Subtitle text="Reveal your culinary art, share your favorite recipe and create gastronomic masterpieces with us." />
+      <SubTitle text="Reveal your culinary art, share your favorite recipe and create gastronomic masterpieces with us." />
       <Button onClick={() => openModal('logout')}>Log Out</Button>
-      <UserInfo/>
+      <UserInfo />
       <RecipePreview />
     </div>
   );
